Update rearrange_cols_rows - Release 1
