+++ include/sudoku.h
@@ -15,3 +15,18 @@ void output_sudoku(int sudoku_solution [ROWS][COLS], int sudoku_puzzle [ROWS][CO
 void create_sudoku_file(int sudoku[ROWS][COLS], char* sudoku_type, int diff_lvl, int output_count);
 
 #endif 
+
+
+/*The below functions are defined only to support the "rearrange_cols_rows" function */
+
+#ifndef REARRANGE_COLS_ROWS_H
+#define REARRANGE_COLS_ROWS_H
+
+int choose_number(int , int );
+void swap(int table[][9]);
+void change_columns(int col_matrix[][9], int p, int q, int r);
+void change_rows(int row_matrix[][9], int x, int y, int z);
+void col_groups(int column_group[][9], int b_1[9][3], int b_2[9][3], int b_3[9][3]);
+void row_groups(int r_group[][9], int c_1[3][9], int c_2[3][9], int c_3[3][9]);
+
+#endif
